structuring locationview
